feat: mobile site nav now shows submenu CTAs, consolidated example items
